Move test data utilities into modal
