+++ packages/builder/src/components/automation/AutomationBuilder/FlowChart/TestDataModal.svelte
@@ -1,19 +1,75 @@
 <script>
   import { tick } from "svelte"
   import { ModalContent, notifications } from "@budibase/bbui"
+  import { AutomationEventType } from "@budibase/types"
   import { automationStore, selectedAutomation } from "@/stores/builder"
   import TestDataEditor from "../../SetupPanel/TestDataEditor.svelte"
   import { cloneDeep } from "lodash/fp"
-  import {
-    buildTriggerSchemaProperties,
-    isTriggerValidForTestData,
-    normalizeParsedJsonForTrigger,
-    parseTestDataForTrigger,
-  } from "../../SetupPanel/testDataUtils"
 
   let failedParse = null
   let trigger = {}
   let schemaProperties = {}
+
+  const rowTriggerEvents = [
+    AutomationEventType.ROW_DELETE,
+    AutomationEventType.ROW_UPDATE,
+    AutomationEventType.ROW_SAVE,
+    AutomationEventType.ROW_ACTION,
+  ]
+
+  const isRowTriggerEvent = event => !!event && rowTriggerEvents.includes(event)
+
+  const getTriggerTableId = target => {
+    const tableId = target?.inputs?.tableId
+    return typeof tableId === "string" ? tableId : undefined
+  }
+
+  const parseTestDataForTrigger = (trigger, data) => {
+    const tableId = getTriggerTableId(trigger)
+    if (isRowTriggerEvent(trigger?.event) && data?.row?.tableId !== tableId) {
+      return {
+        row: { tableId },
+        meta: {},
+        id: "",
+        revision: "",
+      }
+    }
+
+    return cloneDeep(data)
+  }
+
+  const isTriggerValidForTestData = trigger => {
+    if (isRowTriggerEvent(trigger?.event) && !getTriggerTableId(trigger)) {
+      return false
+    }
+    return true
+  }
+
+  const normalizeParsedJsonForTrigger = (trigger, parsedJson) => {
+    const normalized = cloneDeep(parsedJson)
+
+    if (isRowTriggerEvent(trigger?.event)) {
+      const tableId = getTriggerTableId(trigger)
+      if (!normalized.row) {
+        normalized.row = {}
+      }
+      if (normalized.row.tableId !== tableId) {
+        normalized.row.tableId = tableId
+      }
+    }
+
+    return normalized
+  }
+
+  const buildTriggerSchemaProperties = trigger => {
+    let schema = Object.entries(trigger?.schema?.outputs?.properties || {})
+
+    if (trigger?.event === AutomationEventType.APP_TRIGGER) {
+      schema = [["fields", { customType: "fields" }]]
+    }
+
+    return schema
+  }
 
   $: currentTestData = $selectedAutomation.data.testData
